requirements.txt
graphrag==1.0.1
streamlit==1.41.1
theodoretools==0.0.11
python-dotenv~=1.0.1
pyodbc==5.2.0
streamlit-chat==0.1.1
psycopg2-binary==2.9.10
future==1.0.0
langchain-community==0.3.13
openpyxl==3.1.5
pymupdf==1.25.1
pdfplumber==0.11.4
PyPDF2==3.0.1
xlrd==2.0.1
fastapi==0.115.6
uvicorn==0.34.0
fastapi[standard]==0.115.6
pydantic~=2.10.4
requests~=2.32.3
openai==1.58.1
httpx==0.28.1
streamlit-authenticator==0.4.1
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
streamlit-ace==0.1.1
pytest==8.3.4
xlsxwriter==3.2.0
